Ajout de l'instruction LW + Création du mode pas à pas

+++ src/memoryManager.c
@@ -54,7 +54,7 @@ long int loadFromMemory(int index, MainMemory memory){
 	return toReturn;
 }
 
-void storeInMemory(int toStore, int index, MainMemory *memory){
+void storeInMemory(long int toStore, int index, MainMemory *memory){
 
 	//Convertir en binaire le chiffre
 	char binary[32] = {0};
